Revert
This reverts commit c8d620fd5bb75afca6c0de1814c3b0667de0cdb5.

+++ RecurringTask.java
@@ -2,7 +2,7 @@ public class RecurringTask extends Task {
     private int endDate;
     private int frequency;
 
-    RecurringTask(String name, String type, double startTime, double duration, int date, int endDate, int frequency){
+    RecurringTask(String name, String type, float startTime, float duration, int date, int endDate, int frequency){
         super(name,type,startTime,duration,date);
         this.endDate = endDate;
         this.frequency = frequency;
