+++ TransientTask.java
@@ -1,6 +1,6 @@
 public class TransientTask extends Task{
     
-    TransientTask(String name, String type, double startTime, double duration, int date){
+    TransientTask(String name, String type, float startTime, float duration, int date){
         super(name,type,startTime,duration,date);
     }
 }
